experiments/yolov11/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.1
    # via requests
comet-ml==3.48.1
    # via -r requirements.in
configobj==5.0.9
    # via everett
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dulwich==0.22.7
    # via comet-ml
everett==3.1.0
    # via comet-ml
filelock==3.17.0
    # via torch
fonttools==4.55.8
    # via matplotlib
fsspec==2025.2.0
    # via torch
idna==3.10
    # via requests
jinja2==3.1.5
    # via torch
jsonschema==4.23.0
    # via comet-ml
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via
    #   seaborn
    #   ultralytics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.1.1
    # via
    #   contourpy
    #   matplotlib
    #   onnx
    #   opencv-python
    #   pandas
    #   scipy
    #   seaborn
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
onnx==1.17.0
    # via -r requirements.in
opencv-python==4.11.0.86
    # via ultralytics
packaging==24.2
    # via matplotlib
pandas==2.2.3
    # via
    #   seaborn
    #   ultralytics
pillow==11.1.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
protobuf==5.29.3
    # via onnx
psutil==6.1.1
    # via
    #   comet-ml
    #   ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via matplotlib
python-box==6.1.0
    # via comet-ml
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via ultralytics
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   comet-ml
    #   requests-toolbelt
    #   ultralytics
requests-toolbelt==1.0.0
    # via comet-ml
rich==13.9.4
    # via comet-ml
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
scipy==1.15.1
    # via ultralytics
seaborn==0.13.2
    # via ultralytics
semantic-version==2.10.0
    # via comet-ml
sentry-sdk==2.20.0
    # via comet-ml
setuptools==75.8.0
    # via
    #   comet-ml
    #   torch
simplejson==3.19.3
    # via comet-ml
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
torch==2.6.0
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.21.0
    # via ultralytics
tqdm==4.67.1
    # via ultralytics
triton==3.2.0
    # via torch
typing-extensions==4.12.2
    # via
    #   referencing
    #   torch
tzdata==2025.1
    # via pandas
ultralytics==8.3.70
    # via -r requirements.in
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==2.3.0
    # via
    #   comet-ml
    #   dulwich
    #   requests
    #   sentry-sdk
wrapt==1.17.2
    # via comet-ml
wurlitzer==3.1.1
    # via comet-ml
